front-back-garden/requirements.txt
# Front/Back Garden Classifier
# Python 3.11+ recommended for M3 Mac

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0

# Geospatial
osmnx>=1.9.0
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0

# API & networking
requests>=2.31.0

# Visualization
matplotlib>=3.8.0

# Utilities
tqdm>=4.66.0

# API Server
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
httpx>=0.26.0  # Async HTTP client for geocoding
